Extract bundle prelude into separate file
